Inline SearchContext in SearchBox
Apparently, the buffer of a SearchContext is a construct-only property,
so it causes errors when used in GtkBuilder if the buffer is not
properly built at the same time than the context.

Rather, this commit will just inline the SearchContext as an OnceCell
inside the SearchBox itself, so that it's always built after the buffer
has been assigned. Just in case, the editable property of a SearchBox is
now construct-only as well, to make sure that there is always one.

This should remove the Gtk-CRITICAL errors that have been appearing in
the logs since this feature was added.