Remove fuzzy translations
These are auto-generated and uses some guesses to see if they fit.
Unfortunately, I cannot validate whether fuzzy strings are valid, and on
Weblate they don't properly display, so my only chance for safety is to
clear these (unused) translations so that they appear as pending in
Weblate.

Fuzzy translations are not used when the application launches, so this
is just removing the heuristic guesses, not real work.