Report malformed URLs as invalid protocol
If a request such as "localhost:8080/get" is made, the inner parser
treats the URL as "scheme localhost, path 8080/get", which is wrong.

Since this problem is hard to fix, in the meantime I'm using a more
specific error message to report that the URL couldn't be decoded due
to a protocol error.