Prevent request panics from locking the user interface
Previously, if there was panic in a request, because it ran in a
different thread, the execution of the thread was stopped, but the
user interface was left in an inconsistent state.

Now, the panic is catched with catch_unwind(), and if there is an
error during the execution of a request, the user interface is
recovered.

I know that in Rust you are not supposed to catch panics, but if you
are going to leave the user interface in a half-broken state, then
that's unacceptable. Either crash the whole program or let me recover
from the error.

At the same, this commit will fix the error reported at
https://github.com/danirod/cartero/issues/99, which caused a request
to fail if it contained the character 0x00 and was improperly not
marked as binary.